requirements.txt
streamlit
pandas
Pillow
opencv-python-headless
pytesseract
icecream
